Bugfix/3.2/handle topology update (#34)
Do not schedule any new domain discovery while one is already running
If we get an incoming topology update while domain discovery is running, make note of it and schedule it after the domain discovery has finished. Also keep the lock for not issuing new domain discoveries while a domain discovery is running.
This is so that we use as little resources as possible when handling topology updates.