chore(changelog): initial auto tagging and changelog updating (#44)
After this, all commits are required to follow:
https://www.conventionalcommits.org/en/v1.0.0/

The squash message that is created is the same as for the PR, you are
NOT allowed to change this.
Thus create a very good conventional commit message for the PR.
If you change it before squash merging, it will merge it but there is an
extra check that will fail and email that it failed - thus it would then
have to be reverted.